requirements.txt
Flask==1.0.3
torchvision==0.3.0
numpy==1.16.4
Pillow==6.0.0
gunicorn==19.9.0
